chore: email path fix for deployed server
